Remove previous scheduler and provide most of the documentation
