+++ Sources/SpeziScheduler/Schedule/Weekday+Calendar.swift
