+++ Sources/SpeziScheduler/UI/SimpleTile.swift
