+++ Sources/SpeziScheduler/UserInfo/UserInfoKey.swift
