+++ Sources/SpeziSchedulerMacros/UserStorageEntryMacro.swift
@@ -12,6 +12,10 @@ import SwiftSyntax
 import SwiftSyntaxMacros
 
 
+/// The user storage entry macro.
+///
+/// Generates get and set accessors to retrieve the value from the shared repository.
+/// The peer macro generates a `__Key_` prefixed `UserStorageKey`.
 public struct UserStorageEntryMacro {}
 
 
@@ -27,12 +31,19 @@ extension UserStorageEntryMacro: AccessorMacro {
             return [] // diagnostics are provided by the peer macro expansion
         }
 
-        let getAccessor: AccessorDeclSyntax =
-        """
-        get {
-        self[__Key_\(identifier).self]
+        let getAccessor: AccessorDeclSyntax = if let initializer = binding.initializer {
+            """
+            get {
+            self[__Key_\(identifier).self, default: \(initializer.value)]
+            }
+            """
+        } else {
+            """
+            get {
+            self[__Key_\(identifier).self]
+            }
+            """
         }
-        """
 
         let setAccessor: AccessorDeclSyntax =
         """
@@ -51,16 +62,15 @@ extension UserStorageEntryMacro: PeerMacro {
         providingPeersOf declaration: some DeclSyntaxProtocol,
         in context: some MacroExpansionContext
     ) throws -> [DeclSyntax] {
-        // TODO: update name in diagnostics!
         guard let variableDeclaration = declaration.as(VariableDeclSyntax.self) else {
-            throw DiagnosticsError(syntax: declaration, message: "'@UserInfoEntry' can only be applied to a 'var' declaration", id: .invalidSyntax)
+            throw DiagnosticsError(syntax: declaration, message: "'@Property' can only be applied to a 'var' declaration", id: .invalidSyntax)
         }
 
         guard let binding = variableDeclaration.bindings.first,
               variableDeclaration.bindings.count == 1 else {
             throw DiagnosticsError(
                 syntax: declaration,
-                message: "'@UserInfoEntry' can only be applied to a 'var' declaration with a single binding",
+                message: "'@Property' can only be applied to a 'var' declaration with a single binding",
                 id: .invalidSyntax
             )
         }
@@ -68,7 +78,7 @@ extension UserStorageEntryMacro: PeerMacro {
         guard let identifier = binding.pattern.as(IdentifierPatternSyntax.self)?.identifier else {
             throw DiagnosticsError(
                 syntax: declaration,
-                message: "'@UserInfoEntry' can only be applied to a 'var' declaration with a simple name",
+                message: "'@Property' can only be applied to a 'var' declaration with a simple name",
                 id: .invalidSyntax
             )
         }
@@ -81,7 +91,7 @@ extension UserStorageEntryMacro: PeerMacro {
               let extensionDecl = rootContext.as(ExtensionDeclSyntax.self) else {
             throw DiagnosticsError(
                 syntax: declaration,
-                message: "'@UserInfoEntry' can only be applied to 'var' declarations inside of extensions to 'Outcome' or 'ILTask.Context'",
+                message: "'@Property' can only be applied to 'var' declarations inside of extensions to 'Outcome' or 'Task.Context'",
                 id: .invalidSyntax
             )
         }
@@ -94,13 +104,13 @@ extension UserStorageEntryMacro: PeerMacro {
         } else if let memberType = extensionDecl.extendedType.as(MemberTypeSyntax.self),
                   let baseIdentifier = memberType.baseType.as(IdentifierTypeSyntax.self)?.name.identifier?.name,
                   let identifier = memberType.name.identifier?.name,
-                  baseIdentifier == "ILTask", // TODO: do not forget to update
+                  baseIdentifier == "Task",
                   identifier == "Context" {
             keyProtocol = "TaskStorageKey"
         } else {
             throw DiagnosticsError(
                 syntax: declaration,
-                message: "'@UserInfoEntry' can only be applied to 'var' declarations inside of extensions to 'Outcome' or 'ILTask.Context'",
+                message: "'@Property' can only be applied to 'var' declarations inside of extensions to 'Outcome' or 'Task.Context'",
                 id: .invalidSyntax
             )
         }
@@ -111,14 +121,16 @@ extension UserStorageEntryMacro: PeerMacro {
         if let optionalType = typeAnnotation.type.as(OptionalTypeSyntax.self) {
             valueTypeInitializer = optionalType.wrappedType
         } else {
-            // TODO: eventually support with default value?
-            // valueTypeInitializer = typeAnnotation.type
-            // keyProtocol = "TaskStorageKey"
+            valueTypeInitializer = typeAnnotation.type
 
-            throw DiagnosticsError(syntax: binding, message: "Type annotation must use an optional type syntax", id: .invalidSyntax)
+            if binding.initializer == nil {
+                throw DiagnosticsError(
+                    syntax: binding,
+                    message: "A non-optional type requires an initializer expression to provide a default value",
+                    id: .invalidSyntax
+                )
+            }
         }
-
-        // TODO: use the default value provided!
 
         let modifier: TokenSyntax? = variableDeclaration.modifiers
             .compactMap { (modifier: DeclModifierSyntax) -> TokenSyntax? in
